% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{lipidMapsMapping}
\alias{lipidMapsMapping}
\title{SwissLipids API Mapping}
\usage{
lipidMapsMapping(
  context = c("compound", "gene", "protein"),
  input_item,
  input_value,
  output_item
)
}
\arguments{
\item{context}{`character`}

\item{input_item}{`character`}

\item{input_value}{`character`}

\item{output_item}{`character`}
}
\value{
`data.frame`
}
\description{
`lipidMapsMapping` uses the SwissLipids API to perform mapping between
    different database identifier
}
\examples{

lipidMapsMapping(context = "compound",
input_item = "inchi_key",
input_value = "WTJKGGKOPKCXLL-VYOBOKEXSA-N",
output_item = "all")
}
\author{
Michael Witting
}
